Update unitxt and pmt files for new items added for 2025 Christmas Event ; changed skinned item lookup behavior
Because we use item hex as a sort key, the hex lookups of items with names of the form `<name>*` were special-cased
to resolve to the same hex as `<name>` so that e.g. skinned and unskinned Red Rings that are far apart in hex
will sort next to each other.

This was implemented with weapon and Red Ring skins in mind, but skinned mags exist with the same naming convention
but are otherwise not similar to weapon and barrier skins, and this lookup behavior was less useful. Christmas 2025
introduced a new mag `Present*` and a new bug with it: there is no mag named `Present`, but a hex lookup assumes
there is. The special case behavior is now restricted to weapons and barriers. Also in both cases, a hex lookup
will fall back to `<Foo>*` in cases where `<Foo>` doesn't exist, just in case.